fix: corrigido assinatura da função para aceitar como parâmetro int64

+++ internal/domain/user/repository.go
@@ -4,6 +4,6 @@ import "context"
 
 type UserRepository interface {
 	Create(ctx context.Context, user *User, credentials *Credentials) (*User, error)
-	FindByID(ctx context.Context, id int) (*User, error)
+	FindByID(ctx context.Context, id int64) (*User, error)
 	FindByEmail(ctx context.Context, email string) (*User, *Credentials, error)
 }
